Revert "Added more exclusions to checkpoints"
